homework-5 fixed :)

+++ src/main/java/ru/mail/polis/homework/collections/mail/Salary.java
@@ -3,9 +3,9 @@ package ru.mail.polis.homework.collections.mail;
 /**
  * 1 тугрик
  */
-public class Salary extends Mail<Long> {
+public class Salary extends Mail<Integer> {
 
-    public Salary(String recipient, String sender, Long salary) {
+    public Salary(String recipient, String sender, Integer salary) {
         super(recipient, sender, salary);
     }
 }
